requirements.txt
jupyterlab>3.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
elevenlabs==1.12.1
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gradio==5.5.0
gradio_client==1.4.2
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
openai
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rsa==4.9
ruff==0.7.2
safehttpx==0.1.1
safetensors==0.4.5
semantic-version==2.10.0
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
sympy==1.13.1
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
Werkzeug==3.1.2
